Integração com banco de dados

+++ 00_RestAspNet_Scaffold/RestAspNet/Controllers/PersonController.cs
@@ -25,7 +25,7 @@ namespace RestAspNet.Controllers
         }
 
         [HttpGet("{id}")]
-        public IActionResult Get(long id)
+        public IActionResult Get(int id)
         {
             var person = _personService.FindByID(id);
 
@@ -51,7 +51,7 @@ namespace RestAspNet.Controllers
         }
 
         [HttpDelete("{id}")]
-        public IActionResult Delete(long id)
+        public IActionResult Delete(int id)
         {
             _personService.Delete(id);
