+++ 00_RestAspNet_Scaffold/RestAspNet/Services/Implementations/IPersonService.cs
@@ -6,9 +6,9 @@ namespace RestAspNet.Services.Implementations
     public interface IPersonService
     {
         Person Create(Person person);
-        Person FindByID(long id);
+        Person FindByID(int id);
         List<Person> FindAll();
         Person Update(Person person);
-        void Delete(long id);
+        void Delete(int id);
     }
 }
